requirements.txt
pybars3
rich
python-dotenv
rich
flask